requirements.txt
torch
torchvision
opencv-python
pillow
numpy
matplotlib
seaborn
scikit-learn